+++ R/options.R
@@ -111,7 +111,7 @@
 #' @section Advanced:
 #' The following options are likely not needed by a user.
 #' \tabular{lcl}{
-#'   \code{cloudChecksumsFilename} \tab \code{file.path(dirname(.reproducibleTempCacheDir), "checksums.rds")}
+#'   \code{cloudChecksumsFilename} \tab \code{file.path(dirname(.reproducibleTempCacheDir()), "checksums.rds")}
 #'                      \tab Used in \code{\link{cloudCache}} \cr
 #'   \code{length} \tab \code{Inf} \tab Used in \code{\link{Cache}}, specifically to the internal
 #'                      calls to \code{\link{CacheDigest}}. This is passed to \code{digest::digest}.
@@ -124,9 +124,7 @@
 reproducibleOptions <- function() {
   list( # nolint
     reproducible.ask = TRUE,
-    reproducible.cloudChecksumsFilename = normPath(file.path(dirname(.reproducibleTempCacheDir),
-                                                    "checksums.rds")),
-    reproducible.cachePath = normPath(file.path(.reproducibleTempCacheDir)),
+    reproducible.cachePath = normPath(tempdir2("cache")),
     reproducible.cacheSaveFormat = "rds",
     reproducible.conn = NULL,
     reproducible.destinationPath = NULL,
@@ -140,6 +138,7 @@ reproducibleOptions <- function() {
     reproducible.quick = FALSE,
     reproducible.showSimilar = FALSE,
     reproducible.showSimilarDepth = 3,
+    reproducible.tempPath = .reproducibleTempPath(),
     reproducible.useCache = TRUE, # override Cache function
     reproducible.useCloud = FALSE, #
     reproducible.useDBI = TRUE,