+++ R/preProcess.R
@@ -390,6 +390,8 @@ preProcess <- function(targetFile = NULL, url = NULL, archive = NULL, alsoExtrac
     }
   }
   # if it was a nested file
+  browser(expr = exists("._postProcess_8"))
+
   if (any(file_ext(neededFiles) %in% c("zip", "tar", "rar"))) {
     nestedArchives <- .basename(neededFiles[file_ext(neededFiles) %in% c("zip", "tar", "rar")])
     nestedArchives <- normPath(file.path(destinationPath, nestedArchives[1]))
@@ -445,6 +447,7 @@ preProcess <- function(targetFile = NULL, url = NULL, archive = NULL, alsoExtrac
   if (is.null(targetFile) && !is.null(targetFilePath)) {
     targetFile <- .basename(targetFilePath)
   }
+  browser(expr = exists("._postProcess_9"))
 
   ## Convert the fun as character string to function class, if not already
   fun <- .extractFunction(fun)
@@ -484,6 +487,7 @@ preProcess <- function(targetFile = NULL, url = NULL, archive = NULL, alsoExtrac
     on.exit() # remove on.exit because it is done here
   }
 
+  browser(expr = exists("._postProcess_10"))
   failStop <- if (is.null(targetFilePath)) {
     TRUE
   } else if (!isTRUE(file.exists(targetFilePath))) {