+++ tests/testthat/test-filesMissingExtension.R
@@ -5,7 +5,8 @@ test_that("prepInputs works with NULL archive + file without extension, but orig
     testOnExit(testInitOut)
   }, add = TRUE)
   testthat::expect_message(ras <- reproducible::prepInputs(url = "https://github.com/tati-micheletti/host/raw/master/data/unknownExtension",
-                                  alsoExtract = "similar", destinationPath = tempdir()))
+                                  alsoExtract = "similar",
+                                  destinationPath = tempdir2(rndstr(1,6))))
   testthat::expect_is(object = ras, class = "RasterLayer")
 })
 
@@ -15,8 +16,10 @@ test_that("prepInputs WORKS if the file is not originally a .zip, but archive is
   on.exit({
     testOnExit(testInitOut)
   }, add = TRUE)
-  testthat::expect_message(ras <- reproducible::prepInputs(url = "https://github.com/tati-micheletti/host/raw/master/data/unknownTAR",
-                                                         alsoExtract = "similar", archive = "unknownTAR.tar", destinationPath = tempdir()))
+  testthat::expect_message(
+    ras <- reproducible::prepInputs(url = "https://github.com/tati-micheletti/host/raw/master/data/unknownTAR",
+                                    alsoExtract = "similar", archive = "unknownTAR.tar",
+                                    destinationPath = tempdir2(rndstr(1,6))))
   testthat::expect_is(object = ras, class = "RasterLayer")
 })
 
@@ -26,9 +29,10 @@ test_that("prepInputs WORKS if passing archive .zip", {
   on.exit({
     testOnExit(testInitOut)
   }, add = TRUE)
-  testthat::expect_message(ras <- reproducible::prepInputs(url = "https://github.com/tati-micheletti/host/raw/master/data/unknownExtension",
-                                                           archive = "unknownExtension.zip",
-                                                           alsoExtract = "similar", destinationPath = tempdir()))
+  testthat::expect_message(
+    ras <- reproducible::prepInputs(url = "https://github.com/tati-micheletti/host/raw/master/data/unknownExtension",
+                                    archive = "unknownExtension.zip",
+                                    alsoExtract = "similar", destinationPath = tempdir2(rndstr(1,6))))
   testthat::expect_is(object = ras, class = "RasterLayer")
 })
 
@@ -38,8 +42,9 @@ test_that("prepInputs WORKS passing just targetFile that is NOT an archive", {
   on.exit({
     testOnExit(testInitOut)
   }, add = TRUE)
-  testthat::expect_message(ras <- reproducible::prepInputs(url = "https://github.com/tati-micheletti/host/raw/master/data/unknownTIF",
-                                                           alsoExtract = "similar", targetFile = "unknownTIF.tif", destinationPath = tempdir()))
+  testthat::expect_message(
+    ras <- reproducible::prepInputs(url = "https://github.com/tati-micheletti/host/raw/master/data/unknownTIF",
+                                    alsoExtract = "similar", targetFile = "unknownTIF.tif", destinationPath = tempdir2(rndstr(1,6))))
   testthat::expect_is(object = ras, class = "RasterLayer")
 })
 
@@ -49,8 +54,10 @@ test_that("prepInputs WORKS passing archive + targetFile", {
   on.exit({
     testOnExit(testInitOut)
   }, add = TRUE)
-  testthat::expect_message(ras <- reproducible::prepInputs(url = "https://github.com/tati-micheletti/host/raw/master/data/unknownExtension",
-                                                           archive = "unknownExtension.zip", targetFile = "rasterTest.tif",
-                                                           alsoExtract = "similar", destinationPath = tempdir()))
+  testthat::expect_message(
+    ras <- reproducible::prepInputs(url = "https://github.com/tati-micheletti/host/raw/master/data/unknownExtension",
+                                    archive = "unknownExtension.zip", targetFile = "rasterTest.tif",
+                                    alsoExtract = "similar",
+                                    destinationPath = tempdir2(rndstr(1,6))))
   testthat::expect_is(object = ras, class = "RasterLayer")
 })