+++ NEWS.md
@@ -4,6 +4,9 @@ version 1.0.0.9000
 ==============
 
 ## New features
+* Better management of temporary files in package and tests, e.g., during downloading (`preProcess`). Includes 2 new functions, `tempdir2` and `tempfile2` for use with `reproducible` package
+* New option: `reproducible.tempPath`, which is used for the new control of temporary files. Defaults to `file.path(tempdir(), "reproducible")`. This feature was requested to help manage large amounts of 
+temporary objects that were not being easily and automatically cleaned
 * Copying or moving of Cache directories now works after manual moving/copying, followed by `movedCache`.
 * Cache now treats file-backed Rasters as though they had a relative path instead of their absolute path.
   This means that Cache directories can be copied from one location to another and the file-backed `Raster*` will have their filenames updated on the fly during a Cache recovery.